# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Recruita API
  description: >-
    Applicant management, session profiles, health checks, and consent-gated AI
    candidate matching.
  contact:
    name: Recruita Team
    url: https://github.com/mbelraka/recruita
  version: 1.0.0

servers:
  - url: ''
    description: Relative to the SPA origin (dev proxy forwards /api to Spring on port 3001)

tags:
  - name: Health
    description: Operational health checks
  - name: Applicants
    description: Applicant roster CRUD (requires recruita.persistence.enabled=true)
  - name: Profiles
    description: Session profile and privacy consent flags (persistence profile required)
  - name: Match
    description: Score anonymized candidates against a job description
  - name: Smart Action
    description: Parse natural-language recruiter commands into structured actions

paths:
  /api/health:
    get:
      tags: [Health]
      summary: Health check
      operationId: health
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/applicants:
    get:
      tags: [Applicants]
      summary: List applicant summaries
      operationId: listApplicantSummaries
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers:
            ETag:
              schema:
                type: string
            X-Recruita-Roster-Version:
              schema:
                type: integer
                format: int64
            X-Recruita-Roster-Updated-At:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicantSummaryDto'
        '304':
          description: Not Modified
          headers:
            ETag:
              schema:
                type: string
            X-Recruita-Roster-Version:
              schema:
                type: integer
                format: int64
            X-Recruita-Roster-Updated-At:
              schema:
                type: string
                format: date-time
    post:
      tags: [Applicants]
      summary: Create applicant
      operationId: createApplicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveApplicantRequestDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicantDto'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'

  /api/applicants/full:
    get:
      tags: [Applicants]
      summary: List applicants with full fields
      operationId: listApplicantsFull
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicantDto'

  /api/applicants/{id}:
    get:
      tags: [Applicants]
      summary: Get applicant by id
      operationId: getApplicant
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicantDto'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
    put:
      tags: [Applicants]
      summary: Update applicant
      operationId: updateApplicant
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveApplicantRequestDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicantDto'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
    delete:
      tags: [Applicants]
      summary: Delete applicant
      operationId: deleteApplicant
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFoundProblem'

  /api/profiles:
    post:
      tags: [Profiles]
      summary: Create profile
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveProfileRequestDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDto'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'

  /api/profiles/{id}:
    get:
      tags: [Profiles]
      summary: Get profile by id
      operationId: getProfile
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDto'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
    put:
      tags: [Profiles]
      summary: Update profile
      operationId: updateProfile
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveProfileRequestDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDto'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'

  /api/match/cache/invalidate:
    post:
      tags: [Match]
      summary: Invalidate cached match evaluation results
      operationId: invalidateMatchCache
      responses:
        '204':
          description: Cache invalidated

  /api/match:
    post:
      tags: [Match]
      summary: Evaluate match scores
      operationId: match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequestDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '429':
          $ref: '#/components/responses/TooManyRequestsProblem'
        '500':
          $ref: '#/components/responses/InternalErrorProblem'

  /api/match-job:
    post:
      tags: [Match]
      summary: Evaluate match scores (legacy path)
      operationId: matchLegacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRequestDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '429':
          $ref: '#/components/responses/TooManyRequestsProblem'
        '500':
          $ref: '#/components/responses/InternalErrorProblem'

  /api/action/parse:
    post:
      tags: [Smart Action]
      summary: Parse natural-language command
      operationId: parse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseActionCommandRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseActionResponse'
        '500':
          $ref: '#/components/responses/InternalErrorProblem'

components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequestProblem:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblemDetail'
    NotFoundProblem:
      description: Resource or route not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblemDetail'
    ConflictProblem:
      description: Resource conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblemDetail'
    TooManyRequestsProblem:
      description: Rate limit exceeded
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblemDetail'
    InternalErrorProblem:
      description: Server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblemDetail'

  schemas:
    HealthResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    ApplicationStatus:
      type: string
      enum:
        - received
        - screening
        - interview_scheduled
        - shortlisted
        - offer_extended
        - rejected
        - withdrawn

    UiLanguage:
      type: string
      enum: [en, de, fr, it, rm, es]

    ApplicantSummaryDto:
      type: object
      required: [id, skills]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        location:
          type: string
        yearsOfExperience:
          type: number
          format: double
        applicationStatus:
          $ref: '#/components/schemas/ApplicationStatus'
        currentJobTitle:
          type: string
        availableFrom:
          type: string
          format: date
        skills:
          type: array
          items:
            type: string

    SaveApplicantRequestDto:
      allOf:
        - $ref: '#/components/schemas/ApplicantSummaryDto'
        - type: object
          required:
            - name
            - email
            - phone
            - location
            - yearsOfExperience
            - applicationStatus
            - currentJobTitle
          properties:
            notes:
              type: string

    ApplicantDto:
      allOf:
        - $ref: '#/components/schemas/ApplicantSummaryDto'
        - type: object
          properties:
            notes:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    SaveProfileRequestDto:
      type: object
      required:
        - id
        - privacyNoticeAccepted
        - lastLanguage
        - optionalRemoteTranslation
        - optionalGeocoding
        - optionalAiMatching
      properties:
        id:
          type: string
        privacyNoticeAccepted:
          type: boolean
        lastLanguage:
          $ref: '#/components/schemas/UiLanguage'
        optionalRemoteTranslation:
          type: boolean
        optionalGeocoding:
          type: boolean
        optionalAiMatching:
          type: boolean

    ProfileDto:
      allOf:
        - $ref: '#/components/schemas/SaveProfileRequestDto'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    MatchCandidateDto:
      type: object
      required: [id]
      properties:
        id:
          type: string
        skills:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: number
          format: double
        currentJobTitle:
          type: string

    MatchRequestDto:
      type: object
      required: [candidates]
      properties:
        jobDescription:
          type: string
          description: Free-text job description to score against.
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/MatchCandidateDto'
        deterministic:
          type: boolean
          description: When true, uses the built-in deterministic rubric instead of Groq.
        model:
          type: string
        temperature:
          type: number
          format: double
        topP:
          type: number
          format: double
        seed:
          type: integer
          format: int32

    CandidateProfileDto:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
        yearsExperience:
          type: number
          format: double
        topJobTitles:
          type: array
          items:
            type: string
        education:
          type: string

    MatchScoreDto:
      type: object
      properties:
        id:
          type: string
        matchScore:
          type: integer
          format: int32
        matchingSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        candidateProfile:
          $ref: '#/components/schemas/CandidateProfileDto'
        recommendation:
          type: string

    MatchResponseDto:
      type: object
      properties:
        scores:
          type: array
          items:
            $ref: '#/components/schemas/MatchScoreDto'

    ParseActionCommandRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
        language:
          $ref: '#/components/schemas/UiLanguage'

    ActionType:
      type: string
      enum:
        - FILTER_APPLICANTS
        - UPDATE_STATUS
        - EXPORT_DATA
        - CREATE_APPLICANT
        - DELETE_APPLICANT
        - GENERATE_REPORT
        - MATCH_JOB
        - BULK_UPDATE
        - CLARIFY

    FilterParamsDto:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
        minExperience:
          type: number
          format: double
        maxExperience:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        country:
          type: string
        location:
          type: string
        searchTerm:
          type: string

    ClarifyParamsDto:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items:
            type: string

    UpdateStatusParamsDto:
      type: object
      required: [applicantIdentifier, newStatus]
      properties:
        applicantIdentifier:
          type: string
        newStatus:
          $ref: '#/components/schemas/ApplicationStatus'

    ExportFormat:
      type: string
      enum: [csv, json, excel, pdf]

    ExportParamsDto:
      type: object
      required: [format]
      properties:
        format:
          $ref: '#/components/schemas/ExportFormat'
        filters:
          $ref: '#/components/schemas/FilterParamsDto'

    CreateApplicantParamsDto:
      type: object
      required: [name, email, skills, yearsOfExperience, currentJobTitle]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        skills:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: number
          format: double
        currentJobTitle:
          type: string

    DeleteApplicantParamsDto:
      type: object
      required: [applicantIdentifier]
      properties:
        applicantIdentifier:
          type: string

    ReportType:
      type: string
      enum: [pipeline_summary, skills_distribution, experience_analysis]

    GenerateReportParamsDto:
      type: object
      required: [reportType]
      properties:
        reportType:
          $ref: '#/components/schemas/ReportType'
        filters:
          $ref: '#/components/schemas/FilterParamsDto'

    MatchJobParamsDto:
      type: object
      required: [jobDescription]
      properties:
        jobDescription:
          type: string
        limit:
          type: integer
          format: int32

    BulkUpdateUpdatesDto:
      type: object
      properties:
        applicationStatus:
          $ref: '#/components/schemas/ApplicationStatus'
        notes:
          type: string

    BulkUpdateParamsDto:
      type: object
      required: [filters, updates]
      properties:
        filters:
          $ref: '#/components/schemas/FilterParamsDto'
        updates:
          $ref: '#/components/schemas/BulkUpdateUpdatesDto'

    FilterApplicantsActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [FILTER_APPLICANTS]
        params:
          $ref: '#/components/schemas/FilterParamsDto'

    UpdateStatusActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [UPDATE_STATUS]
        params:
          $ref: '#/components/schemas/UpdateStatusParamsDto'

    ExportDataActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [EXPORT_DATA]
        params:
          $ref: '#/components/schemas/ExportParamsDto'

    CreateApplicantActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [CREATE_APPLICANT]
        params:
          $ref: '#/components/schemas/CreateApplicantParamsDto'

    DeleteApplicantActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [DELETE_APPLICANT]
        params:
          $ref: '#/components/schemas/DeleteApplicantParamsDto'

    GenerateReportActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [GENERATE_REPORT]
        params:
          $ref: '#/components/schemas/GenerateReportParamsDto'

    MatchJobActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [MATCH_JOB]
        params:
          $ref: '#/components/schemas/MatchJobParamsDto'

    BulkUpdateActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [BULK_UPDATE]
        params:
          $ref: '#/components/schemas/BulkUpdateParamsDto'

    ClarifyActionDto:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [CLARIFY]
        params:
          $ref: '#/components/schemas/ClarifyParamsDto'

    ParsedActionDto:
      oneOf:
        - $ref: '#/components/schemas/FilterApplicantsActionDto'
        - $ref: '#/components/schemas/UpdateStatusActionDto'
        - $ref: '#/components/schemas/ExportDataActionDto'
        - $ref: '#/components/schemas/CreateApplicantActionDto'
        - $ref: '#/components/schemas/DeleteApplicantActionDto'
        - $ref: '#/components/schemas/GenerateReportActionDto'
        - $ref: '#/components/schemas/MatchJobActionDto'
        - $ref: '#/components/schemas/BulkUpdateActionDto'
        - $ref: '#/components/schemas/ClarifyActionDto'
      discriminator:
        propertyName: type
        mapping:
          FILTER_APPLICANTS: '#/components/schemas/FilterApplicantsActionDto'
          UPDATE_STATUS: '#/components/schemas/UpdateStatusActionDto'
          EXPORT_DATA: '#/components/schemas/ExportDataActionDto'
          CREATE_APPLICANT: '#/components/schemas/CreateApplicantActionDto'
          DELETE_APPLICANT: '#/components/schemas/DeleteApplicantActionDto'
          GENERATE_REPORT: '#/components/schemas/GenerateReportActionDto'
          MATCH_JOB: '#/components/schemas/MatchJobActionDto'
          BULK_UPDATE: '#/components/schemas/BulkUpdateActionDto'
          CLARIFY: '#/components/schemas/ClarifyActionDto'

    ParseActionResponse:
      type: object
      required: [valid, errors]
      properties:
        valid:
          type: boolean
        action:
          $ref: '#/components/schemas/ParsedActionDto'
        errors:
          type: array
          items:
            type: string

    ApiErrorCode:
      type: string
      enum:
        - request-validation
        - invalid-json-body
        - match-validation
        - applicant-not-found
        - applicant-conflict
        - profile-not-found
        - profile-conflict
        - route-not-found
        - match-service-unavailable
        - rate-limit-exceeded
        - internal-error

    ApiProblemDetail:
      description: RFC 7807 problem response with Recruita semantic `code` and legacy `error` fields.
      type: object
      required: [type, title, status, detail, code, error]
      properties:
        type:
          type: string
          format: uri
          example: https://recruita.dev/problems/applicant-not-found
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        error:
          type: string
          description: Legacy duplicate of detail for existing clients.