# backend/src/main/java/com/recruita/api/action/model/ExportParamsDto.java
package com.recruita.api.action.model;

import com.fasterxml.jackson.annotation.JsonInclude;

@JsonInclude(JsonInclude.Include.NON_EMPTY)
public record ExportParamsDto(String format, FilterParamsDto filters) {}