# backend/src/main/java/com/recruita/api/action/model/MatchJobActionDto.java
package com.recruita.api.action.model;

public record MatchJobActionDto(ActionType type, MatchJobParamsDto params)
    implements ParsedActionDto {

  public MatchJobActionDto(MatchJobParamsDto params) {
    this(ActionType.MATCH_JOB, params);
  }
}