# backend/src/main/java/com/recruita/api/generated/api/ApplicantsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.recruita.api.generated.api;

import org.springframework.http.ProblemDetail;
import com.recruita.api.api.dto.applicant.ApplicantDto;
import com.recruita.api.api.dto.applicant.ApplicantSummaryDto;
import java.time.OffsetDateTime;
import com.recruita.api.api.dto.applicant.SaveApplicantRequestDto;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.12.0")
@Validated
public interface ApplicantsApi {

    /**
     * POST /api/applicants : Create applicant
     *
     * @param saveApplicantRequestDto  (required)
     * @return Created (status code 201)
     *         or Invalid request (status code 400)
     *         or Resource conflict (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/applicants",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ApplicantDto> createApplicant(
         @Valid @RequestBody SaveApplicantRequestDto saveApplicantRequestDto
    );


    /**
     * DELETE /api/applicants/{id} : Delete applicant
     *
     * @param id  (required)
     * @return No Content (status code 204)
     *         or Resource or route not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/applicants/{id}",
        produces = { "application/problem+json" }
    )
    
    ResponseEntity<Void> deleteApplicant(
         @PathVariable("id") String id
    );


    /**
     * GET /api/applicants/{id} : Get applicant by id
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Resource or route not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/applicants/{id}",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<ApplicantDto> getApplicant(
         @PathVariable("id") String id
    );


    /**
     * GET /api/applicants : List applicant summaries
     *
     * @param ifNoneMatch  (optional)
     * @return OK (status code 200)
     *         or Not Modified (status code 304)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/applicants",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<ApplicantSummaryDto>> listApplicantSummaries(
         @RequestHeader(value = "If-None-Match", required = false) String ifNoneMatch
    );


    /**
     * GET /api/applicants/full : List applicants with full fields
     *
     * @return OK (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/applicants/full",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<ApplicantDto>> listApplicantsFull(
        
    );


    /**
     * PUT /api/applicants/{id} : Update applicant
     *
     * @param id  (required)
     * @param saveApplicantRequestDto  (required)
     * @return OK (status code 200)
     *         or Invalid request (status code 400)
     *         or Resource or route not found (status code 404)
     *         or Resource conflict (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/applicants/{id}",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ApplicantDto> updateApplicant(
         @PathVariable("id") String id,
         @Valid @RequestBody SaveApplicantRequestDto saveApplicantRequestDto
    );

}