# backend/src/main/java/com/recruita/api/generated/api/MatchApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.recruita.api.generated.api;

import org.springframework.http.ProblemDetail;
import com.recruita.api.api.dto.match.MatchRequestDto;
import com.recruita.api.api.dto.match.MatchResponseDto;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.12.0")
@Validated
public interface MatchApi {

    /**
     * POST /api/match/cache/invalidate : Invalidate cached match evaluation results
     *
     * @return Cache invalidated (status code 204)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/match/cache/invalidate"
    )
    
    ResponseEntity<Void> invalidateMatchCache(
        
    );


    /**
     * POST /api/match : Evaluate match scores
     *
     * @param matchRequestDto  (required)
     * @return OK (status code 200)
     *         or Invalid request (status code 400)
     *         or Rate limit exceeded (status code 429)
     *         or Server error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/match",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<MatchResponseDto> match(
         @Valid @RequestBody MatchRequestDto matchRequestDto
    );


    /**
     * POST /api/match-job : Evaluate match scores (legacy path)
     *
     * @param matchRequestDto  (required)
     * @return OK (status code 200)
     *         or Invalid request (status code 400)
     *         or Rate limit exceeded (status code 429)
     *         or Server error (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/match-job",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<MatchResponseDto> matchLegacy(
         @Valid @RequestBody MatchRequestDto matchRequestDto
    );

}