# backend/src/main/java/com/recruita/api/generated/api/ProfilesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.recruita.api.generated.api;

import org.springframework.http.ProblemDetail;
import com.recruita.api.api.dto.profile.ProfileDto;
import com.recruita.api.api.dto.profile.SaveProfileRequestDto;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.12.0")
@Validated
public interface ProfilesApi {

    /**
     * POST /api/profiles : Create profile
     *
     * @param saveProfileRequestDto  (required)
     * @return Created (status code 201)
     *         or Invalid request (status code 400)
     *         or Resource conflict (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/profiles",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProfileDto> createProfile(
         @Valid @RequestBody SaveProfileRequestDto saveProfileRequestDto
    );


    /**
     * GET /api/profiles/{id} : Get profile by id
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Resource or route not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/profiles/{id}",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<ProfileDto> getProfile(
         @PathVariable("id") String id
    );


    /**
     * PUT /api/profiles/{id} : Update profile
     *
     * @param id  (required)
     * @param saveProfileRequestDto  (required)
     * @return OK (status code 200)
     *         or Invalid request (status code 400)
     *         or Resource or route not found (status code 404)
     *         or Resource conflict (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/profiles/{id}",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProfileDto> updateProfile(
         @PathVariable("id") String id,
         @Valid @RequestBody SaveProfileRequestDto saveProfileRequestDto
    );

}